Move predictors to their own package
